Delete en/androidjava/getting-started/metered-licensing/_index.md
The current implementation of the Metered is not supported in Android. The problem is that it is used https://docs.oracle.com/en/java/javase/17/docs/api/java.xml/javax/xml/XMLConstants.html#FEATURE_SECURE_PROCESSING. This constant is not supported in Android, and we cannot make changes to the Metered classes.